requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pillow==10.4.0
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.2pip
scikit-learn==1.5.1
matplotlib==3.8.4
pydantic==2.7.4
python-multipart==0.0.9
timm==1.0.7
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning==2.3.0
torchmetrics==1.4.0
einops==0.8.0
ultralytics==8.3.23
onnx==1.16.2
onnxruntime==1.18.1
rich==13.7.1
requests==2.32.3